// src/sanity/types.ts
export type SanityImageCrop = {
  _type: "sanity.imageCrop";
  top?: number;
  bottom?: number;
  left?: number;
  right?: number;
};

export type SanityImageHotspot = {
  _type: "sanity.imageHotspot";
  x?: number;
  y?: number;
  height?: number;
  width?: number;
};
// ...
export declare const internalGroqTypeReferenceTo: unique symbol;
// Source: ./src/sanity/queries.ts
// Variable: getPropertiesQuery
// Query: *[_type == "property"]
export type GetPropertiesQueryResult = Array<{
  _id: string;
  _type: "property";
  _createdAt: string;
  _updatedAt: string;
  _rev: string;
  price: number;
  propertyType: string;
  location: string;
  images?: Array<{
    asset?: {
      _ref: string;
      _type: "reference";
      _weak?: boolean;
      [internalGroqTypeReferenceTo]?: "sanity.imageAsset";
    };
    hotspot?: SanityImageHotspot;
    crop?: SanityImageCrop;
    _type: "image";
    _key: string;
  }>;
  bedrooms?: number;
  bathrooms?: number;
  sqft?: number;
  description?: string;
}>;